Correct timer, move timer class position
