+++ mllib-dal/src/main/native/service.h
@@ -42,7 +42,7 @@ using namespace daal::data_management;
 #include <vector>
 
 #include "error_handling.h"
-#include "oneapi/dal/table/detail/csr.hpp"
+#include "oneapi/dal/table/csr.hpp"
 #include "oneapi/dal/table/homogen.hpp"
 
 using namespace oneapi::dal;
